apply threshold to tca within disturbed regions
when labelling individual images, the disturbed region label comes from
the Canadian Forest Service annual map, therefore it is possible that
the image was acquired before the disturbance occurred. In this case it
doesn't make sense to label it as disturbed, since we have no way of
knowing the within year timing of a disturbance apply thershold based on
the mean and standard deviation of the undisturbed forested areas of the
image to the disturbed region to determine if the disturbance has
occurred.